Dry run (do not delete) by default. (#3)
* Dry run (do not delete) by default.

Also, slightly better documentation.

* Make recommended changes.

+++ amnesia.py
@@ -9,6 +9,11 @@ from tweepy import API, Cursor, OAuthHandler
 from tweepy.error import TweepError
 
 
+# By default, actually delete tweet/likes.
+# Set this to True to just print what will be deleted.
+dry_run = twitter_config.get('dry_run', False)
+
+
 if __name__ == "__main__":
     """
     Set up Twitter authentication
@@ -45,7 +50,8 @@ if __name__ == "__main__":
     for status in Cursor(api.user_timeline).items():
         if status.created_at < delete_before:
             try:
-                api.destroy_status(status.id)
+                if not dry_run:
+                    api.destroy_status(status.id)
                 print("Deleted {}".format(status.id))
                 print(status.text.encode('utf-8'))
             except TweepError as err:
@@ -64,7 +70,8 @@ if __name__ == "__main__":
     for status in Cursor(api.favorites).items():
         if status.created_at < delete_before:
             try:
-                api.destroy_favorite(status.id)
+                if not dry_run:
+                    api.destroy_favorite(status.id)
                 print("Un-favorited {}".format(status.id))
                 print(status.text.encode('utf-8'))
             except TweepError as err:
@@ -79,3 +86,5 @@ if __name__ == "__main__":
                 if elapsed_time < FIFTEEN_MINUTES and counter > RATE_LIMIT:
                     time.sleep(FIFTEEN_MINUTES-elapsed_time)
                     counter, elapsed_time, previous_time = 0, 0, 0
+
+
